Promisified setState and fixed snapshot testing
