+++ src/components/ServiceQuestions.js
@@ -10,18 +10,18 @@ const ServiceQuestions = ({
                             last,
                             setPage}) => {
 
-    let previousButton = <button onClick= {(e) => setPage(e, currentPage - 1)}
+    let previousButton = <button onClick= {async (e) => await setPage(e, currentPage - 1)}
                             style={{margin: '2px'}}
                             // &#60; displays "<"
                             type="button" className="btn btn-secondary previous-btn">&#60;</button>
-    let previousPageButton = <button onClick= {(e) => setPage(e, currentPage - 1)}
+    let previousPageButton = <button onClick= {async (e) => await setPage(e, currentPage - 1)}
                                 style={{margin: '2px'}}
                                 type="button" className="btn btn-primary previous-page-btn">{currentPage}</button>
-    let nextButton = <button onClick= {(e) => setPage(e, currentPage + 1)}
+    let nextButton = <button onClick= {async (e) => await setPage(e, currentPage + 1)}
                             style={{margin: '2px'}}
                             // &#62; displays ">"
                             type="button" className="btn btn-secondary next-btn">&#62;</button>
-    let nextPageButton = <button onClick= {(e) => setPage(e, currentPage + 1)}
+    let nextPageButton = <button onClick= {async (e) => await setPage(e, currentPage + 1)}
                                 style={{margin: '2px'}}
                                 type="button" className="btn btn-primary next-page-btn">{currentPage + 2}</button>
 
@@ -53,7 +53,7 @@ const ServiceQuestions = ({
                 }
                 <tr>
                         <td>
-                            <select value={itemCount} onChange={e => setPage(e)}>
+                            <select value={itemCount} onChange={async e => await setPage(e)}>
                                 {
                                     optionValues.map(possibleItemCounts => 
                                         <option key={possibleItemCounts} value={possibleItemCounts}>{possibleItemCounts}</option>)
