+++ src/data/questions.mock.json
@@ -1,86 +1,79 @@
 [
     {
         "id":1,
-        "question":"You get home from work to see your dog sitting around with 6 other dogs playing poker and talking. What do?",
+        "question":"Question1",
         "type":"TrueFalse",
         "choices":"10",
         "serviceAnswers":[]
     },
     {
         "id":2,
-        "question":"The second question!",
+        "question":"Question2",
         "type":"TrueFalse",
-        "choices":"5",
+        "choices":"10",
         "serviceAnswers":[]
     },
     {
         "id":3,
-        "question":"The third question!",
+        "question":"Question3",
         "type":"TrueFalse",
-        "choices":"5",
+        "choices":"10",
         "serviceAnswers":[]
     },
     {
         "id":4,
-        "question":"The fourth question!",
+        "question":"Question4",
         "type":"TrueFalse",
-        "choices":"5",
+        "choices":"10",
         "serviceAnswers":[]
     },
     {
         "id":5,
-        "question":"The fifth question!",
+        "question":"Question5",
         "type":"TrueFalse",
-        "choices":"5",
-        "serviceAnswers":[]
-    },
-    {
-        "id":2,
-        "question":"The second question!",
-        "type":"TrueFalse",
-        "choices":"5",
+        "choices":"10",
         "serviceAnswers":[]
     },
     {
         "id":6,
-        "question":"The sixth question!",
+        "question":"Question6",
         "type":"TrueFalse",
-        "choices":"5",
+        "choices":"10",
         "serviceAnswers":[]
     },
     {
         "id":7,
-        "question":"The seventh question!",
+        "question":"Question7",
         "type":"TrueFalse",
-        "choices":"5",
+        "choices":"10",
         "serviceAnswers":[]
     },
     {
         "id":8,
-        "question":"The eighth question!",
+        "question":"Question8",
         "type":"TrueFalse",
-        "choices":"5",
+        "choices":"10",
         "serviceAnswers":[]
     },
     {
         "id":9,
-        "question":"The ninth question!",
+        "question":"Question9",
         "type":"TrueFalse",
-        "choices":"5",
+        "choices":"10",
         "serviceAnswers":[]
     },
     {
         "id":10,
-        "question":"The tenth question!",
+        "question":"Question10",
         "type":"TrueFalse",
-        "choices":"5",
+        "choices":"10",
         "serviceAnswers":[]
     },
     {
         "id":11,
-        "question":"The eleventh question!",
+        "question":"Question11",
         "type":"TrueFalse",
-        "choices":"5",
+        "choices":"10",
         "serviceAnswers":[]
     }
 ]
