+++ src/test/ServiceQuestions.test.js
@@ -1,11 +1,13 @@
 import React from 'react'
 import ServiceQuestionContainer from '../containers/ServiceQuestionContainer'
 import ServiceQuestionService from '../services/ServiceQuestionService'
+import '../services/ServiceQuestionService.mock'
 import { BrowserRouter } from "react-router-dom";
 import TestRenderer from 'react-test-renderer'
 
 const serviceQuestionService = ServiceQuestionService.getInstance()
-test('Render all service questions correctly', () => {
+
+test('Render all service questions correctly', async () => {
     const testRenderer = TestRenderer.create(
         <BrowserRouter>
             <ServiceQuestionContainer 
@@ -19,10 +21,10 @@ test('Render all service questions correctly', () => {
     expect(tree).toMatchSnapshot()
 
     let testInstance = testRenderer.root
-    let nextPageBtn = testInstance.findByProps({className: "btn btn-primary next-page-btn"})
+    let nextPageBtn = testInstance.findByProps({className: "btn btn-secondary next-btn"})
 
-    nextPageBtn.props.onClick({
-        target: {value: 0}
+    await nextPageBtn.props.onClick({
+        target: {value: 10}
     })
     
     tree = testRenderer.toJSON()
